requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
redis==5.0.1
openai==1.3.3
python-dotenv==1.0.0
tiktoken==0.5.1
numpy==1.26.2
pydantic==2.5.2
jinja2==3.1.2
